backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
nltk
opencv-python
mediapipe
python-multipart
pydantic
numpy
joblib
flask-cors
groq
python-dotenv
    